Add team functionality to API
This commit introduces the team module to the application, allowing users to join and view teams. It includes changes to the app.js file to add the new teams route, as well as the creation of a team model, service, and controller. The user model has also been updated to include a reference to teams.